cli/pyproject.toml
[project]
name = "dina-agent"
version = "0.9.0"
description = "CLI client for Dina Home Node — encrypted memory, PII scrubbing, action gating"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "click==8.3.1",
    "httpx==0.28.1",
    "websockets==15.0.1",
    "cryptography==46.0.6",
    "base58==2.1.1",
    "argon2-cffi==25.1.0",
    "mnemonic==0.21",
    "fastmcp>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
hermes = ["hermes-agent"]  # Optional: enables --runner hermes. Install per Hermes docs.

[project.scripts]
dina = "dina_cli.main:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
